+++ src/DrawableSuits/DrawableSuitsInput.cs
@@ -251,3 +251,4 @@ internal static class DrawableSuitsInput
         }
     }
 }
+